add film-edit component, add update film action

+++ films-data-table/src/app/films/films.component.ts
@@ -35,7 +35,9 @@ export class FilmsComponent implements OnInit {
   }
 
   deleteFilm(id: number): void {
+    if (confirm('Are you sure yu want to delete this film?')) {
     this.store.dispatch(new DeleteFilmAction(id));
+    }
   }
 
   getGenresList(): void {
